DMNotificationObserver has lifetime tied to its owner
Now when the owner is deallocated, any active observers are automatically
invalidated. This works great!

In my first attempt at this, I initially associated the observers with the
owner directly, and they would invalidate on -dealloc. This worked, but
because the notification observers are "public" it was possible to keep
them alive longer than their owners. That is, when an observer received a
notification, there was no guarantee that its owner was still alive. So it
still needed a weak reference to its owner to be safe in this case. But
many classes don't support weak references, so I had a blacklist (taken
from the docs, which must be kept up-to-date), and if the owner was on
that blacklist, DMWeakProxy was used instead. In summary, it was still
ugly.

I solved this by adding an "invalidator" object that's completely private,
so we can ensure that the only reference to it is by its owner.  When the
invalidator gets -dealloc, it calls -invalidate on all owned notification
observers. Even if something else is keeping an observer alive, it will be
invalidated.

Because we are always invalidated when our owner has deallocated, we can
just keep an __unsafe_unretained reference to our owner. No special cases
for stupid weak referencing!

It's still necessary to tread carefully. When we're reacting to our
owner's deallocation, we mustn't cause the owner to be autoreleased, or we
will crash. In the first attempt I had an -owner method that checked if we
had a real weak reference or a proxy, and returned the owner accordingly.
Under ARC, returning a value usually causes it to be retain-autoreleased.
The solution was similar in both cases: If we know our owner is
deallocating, we need to clear our references before they're read.