Added DMKeyValueObserver
This is similar to DMNotificationObserver, but for key-value observing. It
has owner-linked lifetime, like DMNotificationObserver (which has been
factored out into DMAutoInvalidation), and also watches the observation
target to raise an exception if the target is deallocated while the
observer is still active. (I attempted to find a way to have this just
invalidate the observer as well, but haven't been able to; when the target
deallocates it's too late to unregister.)