DMObserverInvalidator calls -invalidate earlier on dealloc
Previously, observers attached to some owner would be sent -invalidate
when the owner deallocates as part of the runtime releasing associated
objects. This occurred after -dealloc had been sent, but before the
object's memory had been freed (see objc_destructInstance).

This worked well and required no method interposing, but the timing of
invalidate was too late for some purposes. For example, key-value
observing through self could not be torn down this way. If an instance
had a KVO observer for 'self.a.b' (that is, observing 'a.b' of object
'self'), at automatic invalidation time 'a' will already have been
released and cleared, so the observation chain can't be torn down. This
leaves dangling KVO references, which can crash. (We don't always get a
warning about it, either.)

These cases required manual (explicit) invalidation, at an earlier point
in time. For the example, -invalidate must be sent sometime before 'a'
is cleared. Note that if a subclass has overridden 'a' (providing its
own getter or backing storage), -invalidate must be send before _its_
dealloc, completely breaking encapsulation. And there was no way to
check if it was happening correctly or not.

Now, when an observer is attached to some owner, the owner's class is
modified to invalidate observers at the beginning of the dealloc
process. This requires adding or replacing the class's -dealloc
implementation. I haven't yet profiled, but I expect the overhead to be
minimal.

This means that DMKeyValueObserver through self (should) work as
expected, phew!