Observer constructor parameter renamed 'attachToOwner:'
This affects DMKeyValueObserver and DMNotificationObserver, and is an
attempt to clarify that the observers will be automatically retained by
the object argument.

I have gone back-and-forth on whether automatic retaining is the best
approach; the alternative is for clients to, say, create an ivar for
each observer to keep it alive. The latter approach is appealing because
it makes the use more explicit, and potentially doesn't require any
dealloc hooks on the client class. But this also implies that when the
last explicit reference goes away (e.g. setting the ivar to nil), the
observer invalidates via dealloc. Unfortunately this behavior isn't
reliable; under ARC, touching the instance in certain ways will extend
its lifetime (typically via a retain/autorelease pair) unintentionally
and non-obviously, so it can outlive its owner. Thus a dealloc hook is
required anyway, as is explicit invalidation. So in summary, this
approach would actually make things harder to debug in the worst way:
observers would 'usually' invalidate at the expected times, except in
some cases where they happened to live a little longer, which would
depend on the compiler's ARC behavior and optimization level (as that
affects ARC output).