Add DMBlockUtilities, currently with one (very useful!) method:
+isObject:implicitlyRetainedByBlock:

which uses private implementation stuff to generate a guess at if an
object is "implicitly" retained by a block. (Capturing a __block
variable is considered "explicit" for this purpose.)

Now I shall add it to appropriate places to detect if we have retain
cycles (which were previously invisible, unless subjected to lengthy and
tedious testing in Instruments or something).