Fix DMBlockUtilities to prevent reading invalid memory
If the end-of-block address was already aligned to sizeof(id), the code
would move too far and check an address out-of-bounds. This led to
intermittent (yet common) false positive reports for the following code:

- (void)setAbstractMedium:(LIManagedAbstractMedium *)abstractMedium;
{
    _abstractMedium = abstractMedium;

    [_coverObserver invalidate];
    _coverObserver = nil;

    if (abstractMedium) {
        NSSet *const depdendentKeys = [[LISceneKitMediumGeometryChildNode mediumGeometryChildNodeSubclassForMedium:abstractMedium] abstractMediumModeledPropertyNamesAffectingGeometry];
        _coverObserver = [DMNotificationObserver observerForName:NSManagedObjectContextObjectsDidChangeNotification object:abstractMedium.managedObjectContext attachedToOwner:self action:^(NSNotification *notification, LICoverImageView *localSelf, DMNotificationObserver *observer) {
            if (![notification.userInfo[NSUpdatedObjectsKey] containsObject:abstractMedium])
                return;
            NSSet *changedKeys = [NSSet setWithArray:abstractMedium.changedValuesForCurrentEvent.allKeys];
            if (![changedKeys intersectsSet:depdendentKeys])
                return;

            localSelf->_cgImage = nil;
            [localSelf setNeedsDisplay:YES];
        }];
    }

    _cgImage = nil;
    [self setNeedsDisplay:YES];
}

The previous version would often incorrectly claim that the action
block above retained 'self'.

I was unable to reproduce this for the tests, and had not seen it in the
other 150 call sites where this was used, so it seems to be a rare, but
repeatable bug.

I've convinced myself that this is now at least more correct, as before
the code would find a captured variable location to check for
NSGlobalBlock instances, which is wrong because NSGlobalBlock captures
no variables by definition.