DMSafeKVC gets methods to detect orphaned dependent key path methods
That is, methods of the form +keyPathsForValuesAffectingFoo, when the
class doesn't have a matching method -foo. The check can be installed
to run on -init; this has some overhead (~0.8%) - more than I'd like -
but we'd only install it on debug builds.

Ideally the tools (i.e. compiler or Xcode) would check this, but this is
better than nothing.