DMKeyValueObserver can observe multiple targets
A common use for DMKeyValueObserver is to observe the same key path of
multiple objects. With Foundation's KVO this is quite lightweight, but
it's slower with DMKeyValueObserver as a full observer instance must be
allocated and linked to the owner (and target, if appropriate); the
action block is -copy'ed each time (requiring pre-copying by the caller
to avoid multiple redundant instances hanging around); and other
overhead.

A single DMKeyValueObserver instance can now observe multiple objects.
The observations are established and torn down using the fast batch
methods on NSArray.

The tricky part is that the observation targets can deallocate
independently of each other, so the DMKeyValueObserver instance may need
to individually remove observations as each target deallocates.
(Alternatively the observer could keep a strong reference to the
targets, but those semantics would deviate from Foundation's KVO.)

To avoid changing the block signature, the current object responsible
for the action firing is in -[DMKeyValueObserver changingObject].
(From scratch it would make more sense to expose it as a block
parameter.)

Results! Here we compare four ways of observing 1000 target objects:

1. Adding a dummy NSObject as an observer repeatedly in a for-in loop

2. Adding a dummy NSObject as an observer to an array of targets using
   NSArray's -addObserver:toObjectsAtIndexes:forKeyPath:options:context:
   method

3. Creating a DMKeyValueObserver for each target in a for-in loop

Relative Speeds (old):
100.0% add/remove toObjectsAtIndexes (batch size 1000)
 83.9% add/remove toObjectsAtIndexes (batch size 100)
 71.0% add/remove toObjectsAtIndexes (batch size 10)
 66.1% for-in loop add/remove (batch size 1000)
 59.6% for-in loop add/remove (batch size 100)
 52.3% for-in loop add/remove (batch size 10)
  9.5% for-in loop DMKeyValueObserver (batch size 1000)
  8.9% for-in loop DMKeyValueObserver (batch size 100)
  8.5% for-in loop DMKeyValueObserver (batch size 10)

So DMKeyValueObserver was about 10x slower than using basic KVO (though
the base methods don't have blocks or invalidation).

Using new DMKeyValueObserver batch observation:

Relative Speeds (new):
100.0% add/remove toObjectsAtIndexes (size 1000)
 85.5% add/remove toObjectsAtIndexes (size 100)
 76.4% DMKeyValueObserver batch (size 1000)
 68.8% add/remove toObjectsAtIndexes (size 10)
 66.8% for-in loop add/remove (size 1000)
 65.2% DMKeyValueObserver batch (size 100)
 60.8% for-in loop add/remove (size 100)
 57.8% for-in loop add/remove (size 10)
 38.5% DMKeyValueObserver batch (size 10)
  5.8% for-in loop DMKeyValueObserver (size 1000)
  5.7% for-in loop DMKeyValueObserver (size 100)
  5.4% for-in loop DMKeyValueObserver (size 10)

A multi-object DMKeyValueObserver is _much_ faster! For an array of 10
objects, a batch observer is 4.5x faster; for 1000 objects, it's 8x.

Note that the single-target case slowed down by about 60% due to
overhead. A subsequent commit will address this.