Create DMManagedObjectObserver class
This encapsulates a common pattern: observing
NSManagedObjectContextObjectsDidChangeNotification, then enumerating the
userInfo dictionary of changes to filter it down to an entity of
interest.

It extends this by taking a set of interesting key paths, which don't
need to be modeled key paths, which are then resolved to their affecting
modeled properties (via +keyPathsForValuesAffectingValueForKey:). In
addition to being affected by modeled properties of the interesting
entity, they can also be affected by properties on related entities,
which are also observed. If a related entity is changed, the inverse
relationships are used to trace back to the managed object of the
interesting entity.