Explicitly copy action blocks when firing for maybe better ARC code
Adding a local reference to the block would optimize away the retain /
release, which would cause a crash if the action callback invalidated
the observer (this is a compiler bug fixed on 10.9, supposedly):

Precise lifetime not respected for locally loaded ivar (Radar 12569789)

Copying the block seems to accomplish the same effect in a cleaner way.
There is also a crash (search FogBugz for
_NSPopoverSendWillCloseNotification) around a notification callback.
Unfortunately we can't reproduce it yet, but maybe this change affects
things.